mpath support 1
started adding support for mpath devs when DMMP is in use.
to-do: change entrypoint in to generateSdEntries to be a selector for
raw paths or mpath managed paths and then ask for array name,
generateSdEntries will call appropriate combination of sdExtract and
getMpathTbl. Also, I should fix all of these func names to be more
consistent.